pyproject.toml
[project]
name = "july"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.17.2",
    "asyncpg==0.30.0",
    "fastapi>=0.124.0",
    "sqlalchemy==2.0.44",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "sqlmodel==0.0.27",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-httpx>=0.36.0",
    "pytest-mock>=3.15.1",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.run]
# addresses this asyncio issue: https://github.com/nedbat/coveragepy/issues/1082
concurrency = ["thread", "greenlet"]
omit = ["july/__main__.py", "july/cli.py"]
source = ["july"]

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
files = "july/**/*.py"
no_implicit_optional = true
python_version = "3.10"
strict_equality = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
addopts = [
    # "--test-alembic",
    "--log-cli-level=warning",
    # "--junitxml=junit/test-results.xml",
    "--cov=july",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
asyncio_mode = "auto"
filterwarnings = [
    # Enable 'error' to treat all warnings as errors
    # "error",
    # The following are all exceptions from asyncpg because the connection pool
    # is not cleared each test. This is okay because it speeds up tests, but if
    # someone has a better answer go for it. There would need to be a teardown
    # action in the clean_database and ensure_test_db_exists fixtures in conftest.py.
    # Honestly, we should probably refactor how databases are initialized and closed
    # in general: https://magicstack.github.io/asyncpg/current/usage.html#connection-pools
    "ignore::ResourceWarning:asyncpg.*",
    "ignore::ResourceWarning",
    "ignore:unclosed transport:ResourceWarning:asyncio.*:",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
norecursedirs = "env/"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = ["env"]
indent-width = 4
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
